refactor: 카테고리 페이지에서 카테고리 관리 컴포넌트 분리
